fix: junit test
