+++ src/main/java/it/gov/pagopa/debtposition/controller/pd/validator/v3/ValidInstallmentsSize.java
@@ -1,0 +1,37 @@
+package it.gov.pagopa.debtposition.controller.pd.validator.v3;
+
+import static java.lang.annotation.ElementType.*;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import javax.validation.Constraint;
+import javax.validation.Payload;
+import javax.validation.constraints.Size;
+
+@Target({METHOD, FIELD, ANNOTATION_TYPE, CONSTRUCTOR, PARAMETER, TYPE_USE})
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+@Constraint(validatedBy = InstallmentsSizeValidator.class)
+public @interface ValidInstallmentsSize {
+  String message() default "Multiple Installment plan not available";
+
+  Class<?>[] groups() default {};
+
+  Class<? extends Payload>[] payload() default {};
+
+  /**
+   * Defines several {@link Size} annotations on the same element.
+   *
+   * @see Size
+   */
+  @Target({METHOD, FIELD, ANNOTATION_TYPE, CONSTRUCTOR, PARAMETER, TYPE_USE})
+  @Retention(RUNTIME)
+  @Documented
+  @interface List {
+
+    ValidInstallmentsSize[] value();
+  }
+}
